Brings together songbird stuff in one file.
