+++ src/player.rs
@@ -1,12 +1,13 @@
-use crate::handlers::TrackErrorNotifier;
 use reqwest::Client as HttpClient;
 use serenity::{
     all::{ChannelId, GuildId},
+    async_trait,
     client::Context,
     prelude::TypeMapKey,
 };
-use songbird::input::YoutubeDl;
-use songbird::{events::TrackEvent, input::Compose};
+use songbird::events::{Event, EventContext, EventHandler as VoiceEventHandler, TrackEvent};
+use songbird::input::{Compose, YoutubeDl};
+use tracing::warn;
 
 pub struct HttpKey;
 
@@ -58,5 +59,24 @@ pub async fn get_yt_track_duration(
     match src.aux_metadata().await {
         Ok(metadata) => metadata.duration,
         Err(_) => None,
+    }
+}
+
+pub struct TrackErrorNotifier;
+
+#[async_trait]
+impl VoiceEventHandler for TrackErrorNotifier {
+    async fn act(&self, ctx: &EventContext<'_>) -> Option<Event> {
+        if let EventContext::Track(track_list) = ctx {
+            for (state, handle) in *track_list {
+                warn!(
+                    "Track {:?} encountered an error: {:?}",
+                    handle.uuid(),
+                    state.playing
+                );
+            }
+        }
+
+        None
     }
 }
